+++ src/kard/users/rewards/client.py
@@ -79,7 +79,7 @@ class RewardsClient:
             CSV list of included resources in the response (e.g "categories"). Allowed value is `categories`.
 
         supported_components : typing.Optional[typing.Union[ComponentType, typing.Sequence[ComponentType]]]
-            UI component types to include in the response. Valid values are shortDescription, longDescription, cta, tags, and detailTags.
+            UI component types to include in the response.
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.
@@ -184,7 +184,7 @@ class RewardsClient:
             CSV list of included resources in the response (e.g "offers,categories"). Allowed values are `offers` and `categories`.
 
         supported_components : typing.Optional[typing.Union[ComponentType, typing.Sequence[ComponentType]]]
-            UI component types to include in included offers. Valid values are shortDescription, longDescription, cta, tags, and detailTags.
+            UI component types to include in included offers.
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.
@@ -295,7 +295,7 @@ class AsyncRewardsClient:
             CSV list of included resources in the response (e.g "categories"). Allowed value is `categories`.
 
         supported_components : typing.Optional[typing.Union[ComponentType, typing.Sequence[ComponentType]]]
-            UI component types to include in the response. Valid values are shortDescription, longDescription, cta, tags, and detailTags.
+            UI component types to include in the response.
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.
@@ -408,7 +408,7 @@ class AsyncRewardsClient:
             CSV list of included resources in the response (e.g "offers,categories"). Allowed values are `offers` and `categories`.
 
         supported_components : typing.Optional[typing.Union[ComponentType, typing.Sequence[ComponentType]]]
-            UI component types to include in included offers. Valid values are shortDescription, longDescription, cta, tags, and detailTags.
+            UI component types to include in included offers.
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.