+++ src/kard/users/rewards/raw_client.py
@@ -77,7 +77,7 @@ class RawRewardsClient:
             CSV list of included resources in the response (e.g "categories"). Allowed value is `categories`.
 
         supported_components : typing.Optional[typing.Union[ComponentType, typing.Sequence[ComponentType]]]
-            UI component types to include in the response. Valid values are shortDescription, longDescription, cta, tags, and detailTags.
+            UI component types to include in the response.
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.
@@ -225,7 +225,7 @@ class RawRewardsClient:
             CSV list of included resources in the response (e.g "offers,categories"). Allowed values are `offers` and `categories`.
 
         supported_components : typing.Optional[typing.Union[ComponentType, typing.Sequence[ComponentType]]]
-            UI component types to include in included offers. Valid values are shortDescription, longDescription, cta, tags, and detailTags.
+            UI component types to include in included offers.
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.
@@ -366,7 +366,7 @@ class AsyncRawRewardsClient:
             CSV list of included resources in the response (e.g "categories"). Allowed value is `categories`.
 
         supported_components : typing.Optional[typing.Union[ComponentType, typing.Sequence[ComponentType]]]
-            UI component types to include in the response. Valid values are shortDescription, longDescription, cta, tags, and detailTags.
+            UI component types to include in the response.
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.
@@ -514,7 +514,7 @@ class AsyncRawRewardsClient:
             CSV list of included resources in the response (e.g "offers,categories"). Allowed values are `offers` and `categories`.
 
         supported_components : typing.Optional[typing.Union[ComponentType, typing.Sequence[ComponentType]]]
-            UI component types to include in included offers. Valid values are shortDescription, longDescription, cta, tags, and detailTags.
+            UI component types to include in included offers.
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.